STY: move InstLibTest
