+++ pysat/tests/instrument_test_class.py
@@ -7,22 +7,16 @@ Can be imported directly for external instrument libraries of pysat instruments.
 
 """
 
-import datetime as dt
-from importlib import import_module
-import tempfile
 import warnings
 
-import pytest
-
-import pysat
-from pysat.utils import generate_instrument_list
+import pysat.tests.classes.cls_instrument_library as cls_inst_lib
 
 
 def initialize_test_inst_and_date(inst_dict):
     """Initialize the instrument object to test and date.
 
-    Paramters
-    ---------
+    Parameters
+    ----------
     inst_dict : dict
         Dictionary containing specific instrument info, generated by
         generate_instrument_list
@@ -36,17 +30,10 @@ def initialize_test_inst_and_date(inst_dict):
 
     """
 
-    test_inst = pysat.Instrument(inst_module=inst_dict['inst_module'],
-                                 tag=inst_dict['tag'],
-                                 inst_id=inst_dict['inst_id'],
-                                 temporary_file_list=True,
-                                 update_files=True)
-    test_dates = inst_dict['inst_module']._test_dates
-    date = test_dates[inst_dict['inst_id']][inst_dict['tag']]
-    return test_inst, date
+    return cls_inst_lib.initialize_test_inst_and_date(inst_dict)
 
 
-class InstTestClass(object):
+class InstTestClass(cls_inst_lib.InstLibTests):
     """Provide standardized tests for pysat instrument libraries.
 
     Note
@@ -62,254 +49,4 @@ class InstTestClass(object):
 
     """
 
-    # Define standard attributes to check.
-    # Needs to be defined here for backwards compatibility.
-    module_attrs = ['platform', 'name', 'tags', 'inst_ids',
-                    'load', 'list_files', 'download']
-    inst_attrs = ['tag', 'inst_id', 'acknowledgements', 'references',
-                  'inst_module']
-    inst_callable = ['load', 'list_files', 'download', 'clean',
-                     'default']
-    attr_types = {'platform': str, 'name': str, 'tags': dict,
-                  'inst_ids': dict, 'tag': str, 'inst_id': str,
-                  'acknowledgements': str, 'references': str}
 
-    def setup_class(self):
-        """Initialize the testing setup once before all tests are run."""
-
-        # Use a temporary directory so that the user's setup is not altered.
-        self.tempdir = tempfile.TemporaryDirectory()
-        self.saved_path = pysat.params['data_dirs']
-        pysat.params['data_dirs'] = self.tempdir.name
-        return
-
-    def teardown_class(self):
-        """Clean up downloaded files and parameters from tests."""
-
-        pysat.params['data_dirs'] = self.saved_path
-        self.tempdir.cleanup()
-        del self.saved_path, self.tempdir
-        return
-
-    def initialize_test_package(self, inst_loc, user_info=None):
-        """Generate custom instrument lists for each category of tests.
-
-        Parameters
-        ----------
-        inst_loc : python subpackage
-            The location of the instrument subpackage to test, e.g.,
-            'pysat.instruments'
-        user_info : dict or NoneType
-            Nested dictionary with user and password info for instrument module
-            name.  If None, no user or password is assumed. (default=None)
-            EX: user_info = {'jro_isr': {'user': 'myname', 'password': 'email'}}
-
-        Returns
-        -------
-        instruments : dict
-            A dictionary containing the lists of instruments from a given
-            package for each category of tests.  The categories are:
-            "names" : A list of all insrument modules by name.
-            "download" : Instrument objects with full download support.
-            "no_download" : Instrument objects without download support.
-
-        """
-
-        # Attach location of package to test object for later reference.
-        self.inst_loc = inst_loc
-
-        # Find all instruments for testing from user-specified location.
-        instruments = generate_instrument_list(inst_loc=inst_loc,
-                                               user_info=user_info)
-
-        # Find all methods in the standard test class.
-        method_list = [func for func in dir(self)
-                       if callable(getattr(self, func))]
-        # Search tests for iteration via pytestmark, update w/ instrument list.
-        for method in method_list:
-            if hasattr(getattr(self, method), 'pytestmark'):
-                # Get list of names of pytestmarks.
-                n_args = len(getattr(self, method).pytestmark)
-                mark_names = [getattr(self, method).pytestmark[j].name
-                              for j in range(0, n_args)]
-                # Add instruments from your library.
-                if 'all_inst' in mark_names:
-                    mark = pytest.mark.parametrize("inst_name",
-                                                   instruments['names'])
-                    getattr(self, method).pytestmark.append(mark)
-                elif 'download' in mark_names:
-                    mark = pytest.mark.parametrize("inst_dict",
-                                                   instruments['download'])
-                    getattr(self, method).pytestmark.append(mark)
-                elif 'no_download' in mark_names:
-                    mark = pytest.mark.parametrize("inst_dict",
-                                                   instruments['no_download'])
-                    getattr(self, method).pytestmark.append(mark)
-
-        return instruments
-
-    def assert_hasattr(self, obj, attr_name):
-        """Provide useful info if object is missing a required attribute."""
-
-        estr = "Object {:} missing attribute {:}".format(obj.__repr__(),
-                                                         attr_name)
-        assert hasattr(obj, attr_name), estr
-        return
-
-    def assert_isinstance(self, obj, obj_type):
-        """Provide useful info if object is the wrong type."""
-
-        estr = "Object {:} is type {:}, but should be type {:}".format(
-            obj.__repr__(), type(obj), obj_type)
-        assert isinstance(obj, obj_type), estr
-        return
-
-    @pytest.mark.all_inst
-    def test_modules_standard(self, inst_name):
-        """Test that modules are importable and have standard properties."""
-
-        # ensure that each module is at minimum importable
-        module = import_module(''.join(('.', inst_name)),
-                               package=self.inst_loc.__name__)
-        # Check for presence of basic instrument module attributes
-        for mattr in self.module_attrs:
-            self.assert_hasattr(module, mattr)
-            if mattr in self.attr_types.keys():
-                self.assert_isinstance(getattr(module, mattr),
-                                       self.attr_types[mattr])
-
-        # Check for presence of required instrument attributes
-        for inst_id in module.inst_ids.keys():
-            for tag in module.inst_ids[inst_id]:
-                inst = pysat.Instrument(inst_module=module, tag=tag,
-                                        inst_id=inst_id)
-
-                # Test to see that the class parameters were passed in
-                self.assert_isinstance(inst, pysat.Instrument)
-                assert inst.platform == module.platform
-                assert inst.name == module.name
-                assert inst.inst_id == inst_id
-                assert inst.tag == tag
-                assert inst.inst_module is not None
-
-                # Test the required class attributes
-                for iattr in self.inst_attrs:
-                    self.assert_hasattr(inst, iattr)
-                    if iattr in self.attr_types:
-                        self.assert_isinstance(getattr(inst, iattr),
-                                               self.attr_types[iattr])
-        return
-
-    @pytest.mark.all_inst
-    def test_standard_function_presence(self, inst_name):
-        """Test that each function is callable, all required functions exist."""
-        module = import_module(''.join(('.', inst_name)),
-                               package=self.inst_loc.__name__)
-
-        # Test for presence of all standard module functions
-        for mcall in self.inst_callable:
-            if hasattr(module, mcall):
-                # If present, must be a callable function
-                assert callable(getattr(module, mcall))
-            else:
-                # If absent, must not be a required function
-                assert mcall not in self.module_attrs
-        return
-
-    @pytest.mark.all_inst
-    def test_instrument_test_dates(self, inst_name):
-        """Test that module has structured test dates correctly."""
-
-        module = import_module(''.join(('.', inst_name)),
-                               package=self.inst_loc.__name__)
-        info = module._test_dates
-        for inst_id in info.keys():
-            for tag in info[inst_id].keys():
-                self.assert_isinstance(info[inst_id][tag], dt.datetime)
-        return
-
-    @pytest.mark.first
-    @pytest.mark.download
-    def test_download(self, inst_dict):
-        """Test that instruments are downloadable."""
-
-        test_inst, date = initialize_test_inst_and_date(inst_dict)
-
-        # check for username
-        dl_dict = inst_dict['user_info'] if 'user_info' in \
-            inst_dict.keys() else {}
-        test_inst.download(date, date, **dl_dict)
-        assert len(test_inst.files.files) > 0
-        return
-
-    @pytest.mark.second
-    @pytest.mark.download
-    @pytest.mark.parametrize("clean_level", ['none', 'dirty', 'dusty', 'clean'])
-    def test_load(self, clean_level, inst_dict):
-        """Test that instruments load at each cleaning level."""
-
-        test_inst, date = initialize_test_inst_and_date(inst_dict)
-        if len(test_inst.files.files) > 0:
-            # Set Clean Level
-            test_inst.clean_level = clean_level
-            target = 'Fake Data to be cleared'
-            test_inst.data = [target]
-            try:
-                test_inst.load(date=date)
-            except ValueError as verr:
-                # Check if instrument is failing due to strict time flag
-                if str(verr).find('Loaded data') > 0:
-                    test_inst.strict_time_flag = False
-                    with warnings.catch_warnings(record=True) as war:
-                        test_inst.load(date=date)
-                    assert len(war) >= 1
-                    categories = [war[j].category for j in range(0, len(war))]
-                    assert UserWarning in categories
-                else:
-                    # If error message does not match, raise error anyway
-                    raise(verr)
-
-            # Make sure fake data is cleared
-            assert target not in test_inst.data
-            # If cleaning not used, something should be in the file
-            # Not used for clean levels since cleaning may remove all data
-            if clean_level == "none":
-                assert not test_inst.empty
-        else:
-            pytest.skip("Download data not available")
-
-        return
-
-    @pytest.mark.download
-    def test_remote_file_list(self, inst_dict):
-        """Test if optional list_remote_files routine exists and is callable."""
-
-        test_inst, date = initialize_test_inst_and_date(inst_dict)
-        name = '_'.join((test_inst.platform, test_inst.name))
-
-        if hasattr(getattr(self.inst_loc, name), 'list_remote_files'):
-            assert callable(test_inst.remote_file_list)
-            # check for username
-            dl_dict = inst_dict['user_info'] if 'user_info' in \
-                inst_dict.keys() else {}
-            files = test_inst.remote_file_list(start=date, stop=date, **dl_dict)
-            # If test date is correctly chosen, files should exist
-            assert len(files) > 0
-        else:
-            pytest.skip("remote_file_list not available")
-
-        return
-
-    @pytest.mark.no_download
-    def test_download_warning(self, inst_dict):
-        """Test that instruments without download support have a warning."""
-
-        test_inst, date = initialize_test_inst_and_date(inst_dict)
-
-        with warnings.catch_warnings(record=True) as war:
-            test_inst.download(date, date)
-
-        assert len(war) >= 1
-        categories = [war[j].category for j in range(0, len(war))]
-        assert UserWarning in categories
-        return
